One way mirror sync

+++ app/src/main/java/com/appmindlab/nano/MirrorWorker.java
@@ -128,31 +128,32 @@ public class MirrorWorker extends Worker {
             /////////////////////////
             // FROM MIRROR (NOTES) //
             /////////////////////////
+            if (!mWorkerParameters.getTags().contains(Const.TO_MIRROR_ONLY_TAG)) {
+                Log.d(Const.TAG, "nano - MirrorWorker: From Mirror ");
 
-            Log.d(Const.TAG, "nano - MirrorWorker: From Mirror ");
+                acquireDataSource();
+                String file_name;
+                for (DocumentFile file : dest_dir.listFiles()) {
+                    // Sanity check
+                    if (file.isDirectory())  continue;
 
-            acquireDataSource();
-            String file_name;
-            for (DocumentFile file : dest_dir.listFiles()) {
-                // Sanity check
-                if (file.isDirectory())  continue;
+                    file_name = file.getName();
+                    if (file_name == null)   continue;
 
-                file_name = file.getName();
-                if (file_name == null)   continue;
+                    if (Arrays.asList(Const.RESERVED_FOLDER_NAMES).contains(file_name)) {
+                        // Notes with reserved folder names need to be removed
+                        file.delete();
+                        continue;
+                    }
 
-                if (Arrays.asList(Const.RESERVED_FOLDER_NAMES).contains(file_name)) {
-                    // Notes with reserved folder names need to be removed
-                    file.delete();
-                    continue;
+                    if (file_name.endsWith(")")) {
+                        // Notes with duplicate names need to be removed
+                        file.delete();
+                        continue;
+                    }
+
+                    importSAFFile(file, false);
                 }
-
-                if (file_name.endsWith(")")) {
-                    // Notes with duplicate names need to be removed
-                    file.delete();
-                    continue;
-                }
-
-                importSAFFile(file, false);
             }
 
             // Do below if not an instant operation
